Move ui/NavBar.js to api/NavBar.js. Clean up main/NavBar.js
