+++ client/main.js
@@ -2,20 +2,13 @@ import React from 'react';
 import { Meteor } from 'meteor/meteor';
 import { render } from 'react-dom';
 import { BrowserRouter, Route, Switch } from 'react-router-dom';
+import { createMuiTheme } from '@material-ui/core/styles';
 import Button from '@material-ui/core/Button';
+import NavBar from '../imports/api/NavBar';
+import { SignInPage } from '../imports/ui/SignInPage';
+import { RegisterPage } from '../imports/ui/RegisterPage';
 import {Editor} from '../imports/ui/Editor.js';
 import { Landing } from '../imports/ui/Landing';
-import { createMuiTheme } from '@material-ui/core/styles';
-
-
-// import AccountsUIWrapper from '../imports/ui/AccountsUIWrapper.js';
-
-import {SignInPage} from '../imports/ui/SignInPage';
-import NavBar from '../imports/ui/NavBar';
-// import {LandingPage} from '../imports/ui/LandingPage';
-import { RegisterPage } from '../imports/ui/RegisterPage';
-// import '../imports/api/Router';
-// import InsetList from '../imports/ui/InsetList';
 
 const theme = createMuiTheme({
   palette: {
@@ -41,20 +34,19 @@ Meteor.startup(() => {
         <div>
 
           <NavBar onSignOut={()=> Meteor.logout(function(error){
+            console.log("Am I even in logout?")
             if(!error) {
               window.location.href = Meteor.absoluteUrl('/signin');
-              // Router.go('/signin');
+              console.log("Signout successfully, but why not redirect???")
+            }else {
+              console.log("why error?", error.reason)
             }
           })}/>
           <Switch>
-
-             <Route path="/signin/" component={SignInPage}/>
-            <Route path="/home/" component={Landing}/>
-            {/* <Route path="/users/:id/" component={LandingPage}/> */}
-            {/* <Route exact path="/" component={SignIn}/> */}
-            {/* <Route path="/signin/" component={SignIn}/> */}
+            <Route path="/signin" component={SignInPage}/>
             <Route path="/register" component={RegisterPage}/>
-            <Route path="/documents/:id/" component={Editor}/>
+            <Route path="/home" component={Landing}/>
+            <Route path="/documents/:id" component={Editor}/>
           </Switch>
         </div>
       </BrowserRouter>
